ultimo actualizado

+++ mi_proyecto/hola_mundo_app/views.py
@@ -11,9 +11,9 @@ def hola_mundo(request):
 
 def sumar(request):
     try:
-        num1 = int(request.GET.get('num1',''))
-        num2 = int(request.GET.get('num2',''))
-        num3 = int(request.GET.get('num3',''))
+        num1 = float(request.GET.get('num1',''))
+        num2 = float(request.GET.get('num2',''))
+        num3 = float(request.GET.get('num3',''))
         suma = num1+num2+num3
         return JsonResponse({'resultado':suma})
     except ValueError:
